[#39] Refactoring package layer
