+++ back-end/src/main/java/com/bangstagram/room/controller/dto/request/RoomSaveRequestDto.java
@@ -1,4 +1,4 @@
-package com.bangstagram.room.controller.dto;
+package com.bangstagram.room.controller.dto.request;
 
 import com.bangstagram.room.domain.model.Room;
 import lombok.Builder;
