+++ back-end/src/main/java/com/bangstagram/timeline/controller/dto/request/TimelineUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.bangstagram.timeline.dto;
+package com.bangstagram.timeline.controller.dto.request;
 
 import lombok.Getter;
