+++ back-end/src/main/java/com/bangstagram/timeline/controller/dto/response/TimelineResponseDto.java
@@ -1,4 +1,4 @@
-package com.bangstagram.timeline.dto;
+package com.bangstagram.timeline.controller.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
