+++ back-end/src/main/java/com/bangstagram/user/service/OAuthKakaoService.java
@@ -1,6 +1,6 @@
 package com.bangstagram.user.service;
 
-import com.bangstagram.user.domain.model.api.response.AuthResponseDto;
+import com.bangstagram.user.controller.dto.response.AuthResponseDto;
 import com.bangstagram.user.domain.model.oauth.kakao.KakaoLoginApi;
 import com.bangstagram.user.domain.model.oauth.kakao.KakaoProfileApi;
 import com.bangstagram.user.util.HttpUtils;
