+++ back-end/src/main/java/com/bangstagram/user/service/UserService.java
@@ -1,9 +1,9 @@
 package com.bangstagram.user.service;
 
-import com.bangstagram.user.domain.model.api.request.AuthRequestDto;
-import com.bangstagram.user.domain.model.api.request.JoinRequestDto;
-import com.bangstagram.user.domain.model.api.response.AuthResponseDto;
-import com.bangstagram.user.domain.model.api.response.JoinResponseDto;
+import com.bangstagram.user.controller.dto.request.AuthRequestDto;
+import com.bangstagram.user.controller.dto.request.JoinRequestDto;
+import com.bangstagram.user.controller.dto.response.AuthResponseDto;
+import com.bangstagram.user.controller.dto.response.JoinResponseDto;
 import com.bangstagram.user.domain.model.user.User;
 import com.bangstagram.user.domain.repository.UserRepository;
 import com.bangstagram.user.security.JWT;
