+++ back-end/src/test/java/com/bangstagram/timeline/service/TimelineServiceTest.java
@@ -1,10 +1,10 @@
 package com.bangstagram.timeline.service;
 
 
+import com.bangstagram.timeline.controller.dto.request.TimelineRequestDto;
+import com.bangstagram.timeline.controller.dto.request.TimelineUpdateRequestDto;
+import com.bangstagram.timeline.controller.dto.response.TimelineResponseDto;
 import com.bangstagram.timeline.domain.repository.TimelineRepository;
-import com.bangstagram.timeline.dto.TimelineRequestDto;
-import com.bangstagram.timeline.dto.TimelineResponseDto;
-import com.bangstagram.timeline.dto.TimelineUpdateRequestDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
