+++ back-end/src/test/java/com/bangstagram/user/service/UserServiceTest.java
@@ -1,7 +1,7 @@
 package com.bangstagram.user.service;
 
-import com.bangstagram.user.domain.model.api.request.JoinRequestDto;
-import com.bangstagram.user.domain.model.api.response.JoinResponseDto;
+import com.bangstagram.user.controller.dto.request.JoinRequestDto;
+import com.bangstagram.user.controller.dto.response.JoinResponseDto;
 import com.bangstagram.user.domain.model.user.User;
 import com.bangstagram.user.domain.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
